size_t vs int warning fix

+++ src/EnergyPlus/WindowManagerExteriorThermal.cc
@@ -406,7 +406,8 @@ namespace WindowManager {
         : m_Surface(surface), m_Window(state.dataSurface->SurfaceWindow(t_SurfNum)), m_ShadePosition(ShadePosition::NoShade), m_SurfNum(t_SurfNum),
           m_SolidLayerIndex(0), m_ConstructionNumber(t_ConstrNum), m_TotLay(getNumOfLayers(state)), m_InteriorBSDFShade(false), m_ExteriorShade(false)
     {
-        if (!state.dataConstruction->Construct(m_ConstructionNumber).WindowTypeBSDF && state.dataSurface->SurfWinShadingFlag.size() >= m_SurfNum) {
+        if (!state.dataConstruction->Construct(m_ConstructionNumber).WindowTypeBSDF &&
+            state.dataSurface->SurfWinShadingFlag.size() >= static_cast<size_t>(m_SurfNum)) {
             if (ANY_SHADE_SCREEN(state.dataSurface->SurfWinShadingFlag(m_SurfNum)) || ANY_BLIND(state.dataSurface->SurfWinShadingFlag(m_SurfNum))) {
                 m_ConstructionNumber = state.dataSurface->SurfWinActiveShadedConstruction(m_SurfNum);
                 m_TotLay = getNumOfLayers(state);
@@ -490,7 +491,8 @@ namespace WindowManager {
         auto ConstrNum = m_ConstructionNumber;
 
         // BSDF window do not have special shading flag
-        if (!state.dataConstruction->Construct(ConstrNum).WindowTypeBSDF && state.dataSurface->SurfWinShadingFlag.size() >= m_SurfNum) {
+        if (!state.dataConstruction->Construct(ConstrNum).WindowTypeBSDF &&
+            state.dataSurface->SurfWinShadingFlag.size() >= static_cast<size_t>(m_SurfNum)) {
             if (ANY_SHADE_SCREEN(state.dataSurface->SurfWinShadingFlag(m_SurfNum)) || ANY_BLIND(state.dataSurface->SurfWinShadingFlag(m_SurfNum))) {
                 ConstrNum = state.dataSurface->SurfWinActiveShadedConstruction(m_SurfNum);
             }
